+++ packages/app-config-writer/src/preview-config/ui5-yaml.ts
@@ -1,5 +1,5 @@
 import { basename, join } from 'path';
-import { createPreviewMiddlewareConfig } from '../variants-config/ui5-yaml';
+import { createPreviewMiddlewareConfig } from '../common/ui5-yaml';
 import { ensurePreviewMiddlewareDependency, extractUrlDetails, isValidPreviewScript } from './package-json';
 import { FileName, getAllUi5YamlFileNames, getWebappPath, readUi5Yaml, type Package } from '@sap-ux/project-access';
 import { getPreviewMiddleware, isFioriToolsDeprecatedPreviewConfig } from '../common/utils';
